refactor(web): relocate the web server to lib/web/index.ts
  lib/servers/web.ts -> lib/web/index.ts   (servers/ removed)
  test/servers/      -> test/web/

web.ts's own imports are unchanged (web/ is the same depth under lib/ as
servers/ was); all 10 consumers updated to the new path.
