+++ benches/server.bench.ts
@@ -6,7 +6,7 @@
  * it is affected by OS TCP TIME_WAIT state accumulated by the test suite that runs
  * before bench-check in `make release`, making it unreliable as a regression gate.
  */
-import { HTTPServer } from "../lib/servers/web.ts";
+import { HTTPServer } from "../lib/web/index.ts";
 
 Deno.bench("server: create HTTPServer instance", {
   group: "server",
