+++ test/index.ts
@@ -2,5 +2,5 @@ import './setup.ts';
 import './commands/index.ts';
 import './flags/index.ts';
 import './inputs/index.ts';
-import './servers/index.ts';
+import './web/index.ts';
 import './setup/index.ts';
