+++ test/web/http-test.ts
@@ -3,7 +3,7 @@ import http from 'node:http';
 import net from 'node:net';
 // @deno-types="npm:@types/ws"
 import WebSocket from 'ws';
-import { HTTPServer } from '../../lib/servers/web.ts';
+import { HTTPServer } from '../../lib/web/index.ts';
 import { bindServerToPort } from '../../lib/setup/bind-server-to-port.ts';
 
 module('Servers | bindServerToPort | port selection', { concurrency: true }, () => {
